Check argument quality

Check argument quality

checkArguments(
  dp,
  tz,
  minimum.detections,
  max.interval,
  speed.method = c("last to first", "last to last"),
  speed.warning,
  speed.error,
  start.time,
  stop.time,
  report,
  auto.open,
  save.detections,
  jump.warning,
  jump.error,
  inactive.warning,
  inactive.error,
  exclude.tags,
  override,
  print.releases,
  plot.detections.by = c("stations", "arrays"),
  if.last.skip.section = NULL,
  replicates = NULL,
  section.minimum = NULL,
  section.order = NULL,
  timestep = c("days", "hours")
)

Arguments

dp
A preloaded datapack (or NULL if no data was preloaded).

tz
The time zone of the study area. Must match one of the values present in timezones.

minimum.detections
For tags with only one movement event, defines the minimum number of times a tag must have been recorded during the study period for it to be considered true detections and not random noise. Defaults to 2.

max.interval
The number of minutes that must pass between detections for a new event to be created. Defaults to 60.

speed.method
Can take two forms: 'last to first' or 'last to last'. If 'last to first' (default), the last detection on the previous array is matched to the first detection on the target array to perform the calculations. If 'last to last', the last detection on ´the previous array is matched to the last detection on the target array to perform the calculations.

speed.warning
If a tag moves at a speed equal or greater than speed.warning (in metres per second), a warning is issued. If left NULL (default), no warnings are issued.

speed.error
If a tag moves at a speed equal or greater than speed.error (in metres per second), user intervention is suggested. If left NULL (default), user intervention is never suggested.

start.time
Detection data prior to the timestamp set in start.time (in YYYY-MM-DD HH:MM:SS format) is not considered during the analysis.

stop.time
Detection data posterior to the timestamp set in stop.time (in YYYY-MM-DD HH:MM:SS format) is not considered during the analysis.

report
Logical. Should an HTML report be created at the end of the analysis? NOTE: Setting report to TRUE will generate an HTML file in the current directory. Additionally, if auto.open = TRUE (default), the web browser will automatically be launched to open the report once the function terminates.

auto.open
Logical: Should the report be automatically opened once the analysis is over? Defaults to TRUE. NOTE: If report = TRUE and auto.open = TRUE, the web browser will automatically be launched to open the report once the function terminates.

save.detections
Logical: Should the processed detections be saved for future runs?

jump.warning
If a tag crosses a number of arrays equal or greater than jump.error without being detected, a warning is issued. If left NULL (default), no warnings are issued.

jump.error
If a tag crosses a number of arrays equal or greater than jump.error without being detected, user intervention is suggested. If left NULL (default), user intervention is never suggested.

inactive.warning
If a tag spends a number of days equal or greater than inactive.error in a given array at the tail of the respective detections, a warning is issued. If left NULL (default), no warnings are issued.

inactive.error
If a tag spends a number of days equal or greater than inactive.error in a given array at the tail of the respective detections, user intervention is suggested. If left NULL (default), user intervention is never suggested.

exclude.tags
A vector of tags that should be excluded from the detection data before any analyses are performed. Intended to be used if stray tags from a different code space but with the same signal as a target tag are detected in the study area.

override
A vector of signals for which the user intends to manually define which movement events are valid and invalid.

print.releases
Logical: Should the release sites be printed in the study area diagrams?

plot.detections.by
The type of y axis desired for the individual detection plots. One of "stations" (default) or "arrays".

if.last.skip.section
Logical: Should a tag detected at the last array of a given section be considered to have disappeared in the next section?

replicates
A list containing, for each array to which intra-array efficiency is to be calculated: The standard names of the stations to be used as a replicate. See the vignettes for more details.

section.minimum
If a tag has less than section.minimum consecutive detections in a section, a warning is issued. Defaults to 2.

section.order
A vector containing the order by which sections should be aligned in the results.

timestep
The resolution desired for the residency calculations. One of "days" (default) or "hours".

Value

updated parameters