Calculate Distances Matrix

Using a previously created transition layer (see transitionLayer), calculates the distances between spatial points. Adapted from Grant Adams' script "distance to closest mpa". if the argument 'actel' is set to TRUE (default), an actel-compatible matrix is generated, and the user will be asked if they would like to save the matrix as 'distances.csv' in the current directory.

distancesMatrix(
  t.layer,
  starters = NULL,
  targets = starters,
  coord.x = "x",
  coord.y = "y",
  id.col = NULL,
  actel = TRUE
)

Arguments

t.layer
A TransitionLayer object, generated by transitionLayer.

starters
A data frame with the points from which to start measuring the distance. Ignored if actel = TRUE (default), as the 'spatial.csv' is loaded as starters.

targets
A data frame with the points to which a way must be found. Ignored if actel = TRUE (default), as the 'spatial.csv' is loaded as targets.

coord.x, coord.y
The names of the columns containing the x and y coordinates in the starters and targets. Must be identical in the starters and targets.

id.col
The name of the column containing the IDs of the points to be used as starters and targets. Must be identical in both files. Ignored if actel = TRUE (default), as the stations' standard names are used.

actel
Logical: Should the distance matrix be optimized for actel? Defaults to TRUE.

Value

A matrix with the distances between each pair of points.

Details

It is highly recommended to read the vignette regarding distances matrix before running this function. You can find it by running vignette('a-2_distances_matrix', 'actel') or browseVignettes('actel')

Examples

# \donttest{
# check if R can run the distance functions
aux <- c(
  length(suppressWarnings(packageDescription("raster"))),
  length(suppressWarnings(packageDescription("gdistance"))),
  length(suppressWarnings(packageDescription("sp"))),
  length(suppressWarnings(packageDescription("tools"))),
  length(suppressWarnings(packageDescription("rgdal"))))
missing.packages <- sapply(aux, function(x) x == 1)

if (any(missing.packages)) {
  message("Sorry, this function requires packages '",
    paste(c("raster", "gdistance", "sp", "tools", "rgdal")[missing.packages], collapse = "', '"),
    "' to operate. Please install ", ifelse(sum(missing.packages) > 1, "them", "it"),
    " before proceeding.")
} else {
  if (suppressWarnings(require("rgdal"))) {
    # move to a temporary directory
    old.wd <- getwd()
    setwd(tempdir())

    # Fetch location of actel's example files
    aux <- system.file(package = "actel")[1]

    # create a temporary spatial.csv file
    file.copy(paste0(aux, "/example_spatial.csv"), "spatial.csv")

    # import the shape file and use the spatial.csv
    # to check the extents.
    x <- loadShape(path = aux, shape = "example_shapefile.shp",
      coord.x = "x", coord.y = "y", size = 20)

    # Build the transition layer
    t.layer <- transitionLayer(x)

    # compile the distances matrix. Columns x and y in the spatial dataframe
    # contain the coordinates of the stations and release sites.
    distancesMatrix(t.layer, coord.x = 'x', coord.y = 'y')

    # return to original directory
    setwd(old.wd)
    rm(old.wd)
  } else {
    message("Sorry, it appears that rgdal is not being able to load.")
  }
}

#> Loading required package: rgdal

#> Loading required package: sp

#> rgdal: version: 1.5-16, (SVN revision 1050)
#> Geospatial Data Abstraction Library extensions to R successfully loaded
#> Loaded GDAL runtime: GDAL 3.0.4, released 2020/01/28
#> Path to GDAL shared files: C:/Program Files/R/R-4.0.2/library/rgdal/gdal
#> GDAL binary built with GEOS: TRUE 
#> Loaded PROJ runtime: Rel. 6.3.1, February 10th, 2020, [PJ_VERSION: 631]
#> Path to PROJ shared files: C:/Program Files/R/R-4.0.2/library/rgdal/proj
#> Linking to sp version:1.4-2
#> To mute warnings of possible GDAL/OSR exportToProj4() degradation,
#> use options("rgdal_show_exportToProj4_warnings"="none") before loading rgdal.

#> M: Extending the shape ranges with open water to ensure the stations fit inside it.

#> M: Applying a small correction to the shape extent to ensure an integer number of pixels.

#> M: New shape extent:

#>         min       max
#> x  445488.1  477088.1
#> y 6242151.8 6247911.8

#> M: Chosen pixel size: 20 
#> M: Resulting pixel dimensions:

#>    x    y 
#> 1580  288 

#> M: Burning the shape into a raster. This process may take several minutes depending on the shape size and chosen pixel size.

#> M: Constructing the transition layer. This process may take several minutes depending on the study area size and chosen pixel size.

#> M: Creating actel-compatible distances matrix.

rm(aux, missing.packages)
# }