Calculate estimated last-array efficiency

Calculate estimated last-array efficiency

dualArrayCJS(input)

Arguments

input
A presence/absence matrix.

Value

A list containing:

absolutes A matrix with the absolute number of tags detected at each replicate and at both,

single.efficiency A vector of calculated array detection efficiencies for each of the replicates,

combined.efficiency The value of the combined detection efficiency for the array.

References

Perry et al (2012), 'Using mark-recapture models to estimate survival from telemetry data'. url: https://www.researchgate.net/publication/256443823_Using_mark-recapture_models_to_estimate_survival_from_telemetry_data

Examples

# prepare a dummy presence/absence matrix
x <- matrix(c(TRUE, TRUE, TRUE, TRUE, FALSE, TRUE), ncol = 2)
colnames(x) <- c("R1", "R2")

# run CJS
dualArrayCJS(x)

#> $absolutes
#>                     
#> detected at R1:    3
#> detected at R2:    2
#> detected at both:  2
#> 
#> $single.efficiency
#>        R1        R2 
#> 1.0000000 0.6666667 
#> 
#> $combined.efficiency
#> [1] 1
#>