Extract speeds from the analysis results.

Extract speeds from the analysis results.

getSpeeds(
  input,
  type = c("all", "forward", "backward"),
  direct = FALSE,
  n.events = c("first", "all", "last")
)

Arguments

input
An actel results object generated by explore, migration or residency.

type
The type of movements to record. One of "all", "forward", or "backward". In the two last options, only the forward or backwards (relatively to the study area structure) movement speeds are returned.

direct
Logical: Extract only speeds between arrays that are directly connected (i.e. neighbouring arrays)?

n.events
The events to record. One of "first", "all", or "last".

Value

A data frame with the following columns:

Tag: The tag of the animal who performed the recorded speed

Event: The valid event where the speed was recorded

From.array: The array from which the tags left

From.station: The station from which the tags left

To.array: The array to which the tags arrived

To.station: The station to which the tags arrived

Speed: The speed recorded in the described movement

Examples

# using the example results loaded with actel
getSpeeds(example.results)

#>            Tag Event From.array From.station To.array To.station    Speed
#>   1: R64K-4451     2         A1         St.2       A2       St.3 0.547642
#>   2: R64K-4451     3         A2         St.3       A3       St.4 0.159720
#>   3: R64K-4451     4         A3         St.4       A4       St.5 0.444482
#>   4: R64K-4451     5         A4         St.5       A5       St.6 2.120419
#>   5: R64K-4451     6         A5         St.6       A6       St.7 0.857434
#>  ---                                                                     
#> 443: R64K-4549     5         A4         St.5       A5       St.6 2.892857
#> 444: R64K-4549     6         A5         St.6       A6       St.7 0.633083
#> 445: R64K-4549     7         A6         St.7       A7      St.11 0.028322
#> 446: R64K-4549    13         A7        St.10       A8      St.13 0.031147
#> 447: R64K-4549     1    Release          RS1       A1       St.2 0.001575

# You can specify which direction of movement to extract with 'type'
getSpeeds(example.results, type = "forward")

#>            Tag Event From.array From.station To.array To.station    Speed
#>   1: R64K-4451     2         A1         St.2       A2       St.3 0.547642
#>   2: R64K-4451     3         A2         St.3       A3       St.4 0.159720
#>   3: R64K-4451     4         A3         St.4       A4       St.5 0.444482
#>   4: R64K-4451     5         A4         St.5       A5       St.6 2.120419
#>   5: R64K-4451     6         A5         St.6       A6       St.7 0.857434
#>  ---                                                                     
#> 440: R64K-4549     5         A4         St.5       A5       St.6 2.892857
#> 441: R64K-4549     6         A5         St.6       A6       St.7 0.633083
#> 442: R64K-4549     7         A6         St.7       A7      St.11 0.028322
#> 443: R64K-4549    13         A7        St.10       A8      St.13 0.031147
#> 444: R64K-4549     1    Release          RS1       A1       St.2 0.001575

# or
getSpeeds(example.results, type = "backward")

#>          Tag Event From.array From.station To.array To.station     Speed
#> 1: R64K-4508    17         A9        St.15       A8      St.14  0.758962
#> 2: R64K-4526    17         A9        St.15       A8      St.14  0.317149
#> 3: R64K-4543     6         A5         St.6       A4       St.5 15.576923

# and also how many events per tag (this won't change the output 
# with the example.results, only because these results are minimal).
getSpeeds(example.results, n.events = "first")

#>            Tag Event From.array From.station To.array To.station    Speed
#>   1: R64K-4451     2         A1         St.2       A2       St.3 0.547642
#>   2: R64K-4451     3         A2         St.3       A3       St.4 0.159720
#>   3: R64K-4451     4         A3         St.4       A4       St.5 0.444482
#>   4: R64K-4451     5         A4         St.5       A5       St.6 2.120419
#>   5: R64K-4451     6         A5         St.6       A6       St.7 0.857434
#>  ---                                                                     
#> 443: R64K-4549     5         A4         St.5       A5       St.6 2.892857
#> 444: R64K-4549     6         A5         St.6       A6       St.7 0.633083
#> 445: R64K-4549     7         A6         St.7       A7      St.11 0.028322
#> 446: R64K-4549    13         A7        St.10       A8      St.13 0.031147
#> 447: R64K-4549     1    Release          RS1       A1       St.2 0.001575

# or
getSpeeds(example.results, n.events = "all")

#>            Tag Event From.array From.station To.array To.station    Speed
#>   1: R64K-4451     1    Release          RS1       A1       St.2 0.001987
#>   2: R64K-4451     2         A1         St.2       A2       St.3 0.547642
#>   3: R64K-4451     3         A2         St.3       A3       St.4 0.159720
#>   4: R64K-4451     4         A3         St.4       A4       St.5 0.444482
#>   5: R64K-4451     5         A4         St.5       A5       St.6 2.120419
#>  ---                                                                     
#> 448: R64K-4549     4         A3         St.4       A4       St.5 0.150715
#> 449: R64K-4549     5         A4         St.5       A5       St.6 2.892857
#> 450: R64K-4549     6         A5         St.6       A6       St.7 0.633083
#> 451: R64K-4549     7         A6         St.7       A7      St.11 0.028322
#> 452: R64K-4549    13         A7        St.10       A8      St.13 0.031147

# or
getSpeeds(example.results, n.events = "last")

#>            Tag Event From.array From.station To.array To.station    Speed
#>   1: R64K-4451     2         A1         St.2       A2       St.3 0.547642
#>   2: R64K-4451     3         A2         St.3       A3       St.4 0.159720
#>   3: R64K-4451     4         A3         St.4       A4       St.5 0.444482
#>   4: R64K-4451     5         A4         St.5       A5       St.6 2.120419
#>   5: R64K-4451     6         A5         St.6       A6       St.7 0.857434
#>  ---                                                                     
#> 443: R64K-4549     5         A4         St.5       A5       St.6 2.892857
#> 444: R64K-4549     6         A5         St.6       A6       St.7 0.633083
#> 445: R64K-4549     7         A6         St.7       A7      St.11 0.028322
#> 446: R64K-4549    13         A7        St.10       A8      St.13 0.031147
#> 447: R64K-4549     1    Release          RS1       A1       St.2 0.001575