Load Spatial File

Loads a spatial file prepared for actel and appends the Standard.name column. Additionally, performs a series of quality checks on the contents of the target file.

loadSpatial(input = "spatial.csv", section.order = NULL)

Arguments

input
Either a data frame or the name of an input file with spatial data in the actel format.

section.order
A vector containing the order by which sections should be aligned in the results.

Value

A data frame with the spatial information present in 'spatial.csv' and the Standard.name column.

Examples

# This function requires the presence of a file with spatial information

# Fetch location of actel's example files
aux <- system.file(package = "actel")[1]

# run loadSpatial on the temporary spatial.csv file
loadSpatial(input = paste0(aux, '/example_spatial.csv'))

#>    Station.name Latitude Longitude        x       y Array Section       Type
#> 1    Station -1 56.35474  8.629019 477075.2 6245625    A0   River Hydrophone
#> 2     Station 1 56.35672  8.603097 475474.6 6245854    A1   River Hydrophone
#> 3     Station 2 56.36505  8.510094 469734.4 6246818    A2   River Hydrophone
#> 4     Station 3 56.34976  8.385683 462033.4 6245178    A3   River Hydrophone
#> 5     Station 4 56.32894  8.309133 457279.2 6242906    A4   River Hydrophone
#> 6     Station 5 56.32576  8.306444 457109.4 6242553    A5   River Hydrophone
#> 7     Station 6 56.32267  8.303267 456909.4 6242211    A6   River Hydrophone
#> 8     Station 7 56.32865  8.249125 453568.3 6242912    A7   Fjord Hydrophone
#> 9     Station 8 56.32610  8.246447 453399.6 6242630    A7   Fjord Hydrophone
#> 10    Station 9 56.32388  8.240797 453047.5 6242387    A7   Fjord Hydrophone
#> 11   Station 10 56.32192  8.239672 452975.5 6242169    A7   Fjord Hydrophone
#> 12   Station 11 56.37173  8.132342 446407.4 6247792    A8   Fjord Hydrophone
#> 13   Station 12 56.37221  8.130017 446264.5 6247847    A8   Fjord Hydrophone
#> 14   Station 13 56.37250  8.127761 446125.6 6247881    A8   Fjord Hydrophone
#> 15   Station 14 56.37199  8.118439 445549.1 6247831    A9     Sea Hydrophone
#> 16   Station 15 56.37256  8.118169 445533.2 6247895    A9     Sea Hydrophone
#> 17   Station 16 56.37200  8.117658 445500.9 6247833    A9     Sea Hydrophone
#> 18          RS1 56.34900  8.620400 476539.0 6244988    A1    <NA>    Release
#>    Standard.name
#> 1           St.1
#> 2           St.2
#> 3           St.3
#> 4           St.4
#> 5           St.5
#> 6           St.6
#> 7           St.7
#> 8           St.8
#> 9           St.9
#> 10         St.10
#> 11         St.11
#> 12         St.12
#> 13         St.13
#> 14         St.14
#> 15         St.15
#> 16         St.16
#> 17         St.17
#> 18           RS1