Load, structure and check the inputs

Load, structure and check the inputs

loadStudyData(
  tz,
  override = NULL,
  start.time,
  stop.time,
  save.detections = FALSE,
  section.order = NULL,
  exclude.tags,
  disregard.parallels = TRUE,
  discard.orphans = FALSE
)

Arguments

tz
The time zone of the study area. Must match one of the values present in timezones.

override
A vector of signals for which the user intends to manually define which movement events are valid and invalid.

start.time
Detection data prior to the timestamp set in start.time (in YYYY-MM-DD HH:MM:SS format) is not considered during the analysis.

stop.time
Detection data posterior to the timestamp set in stop.time (in YYYY-MM-DD HH:MM:SS format) is not considered during the analysis.

save.detections
Logical: Should the processed detections be saved for future runs?

section.order
A vector containing the order by which sections should be aligned in the results.

exclude.tags
A vector of tags that should be excluded from the detection data before any analyses are performed. Intended to be used if stray tags from a different code space but with the same signal as a target tag are detected in the study area.

disregard.parallels
Logical: Should the presence of parallel arrays invalidate potential efficiency peers? See the vignettes for more details.

discard.orphans
Logical: Should actel automatically discard detections that do not fall within receiver deployment periods, or that were recorded before the respective animals were released?

Value

A list containing:

bio: A data frame corresponding to 'biometrics.csv'

sections: A vector of the study area sections (or NULL in the case of the explore function)

deployments: A list corresponding to 'deployments.csv'

spatial: A list corresponding to 'spatial.csv'

dot: A data frame containing the connections between arrays

arrays: A list containing detailed information on the arrays

dotmat: A matrix of the distance (in number of arrays) between pairs of arrays

dist.mat: A matrix of the distances (in metres) between stations (if a 'distances.csv' is present)

detections.list: A list containing the detection data for each tag

paths: A list of the all array paths between each pair of arrays.