Migration Analysis

The migration analysis runs the same initial checks as explore, but on top of it, it analyses the animal behaviour. By selecting the arrays that lead to success, you can define whether or not your animals survived the migration. Additional plots help you find out if some animal/tag has been acting odd. Multiple options allow you to tweak the analysis to fit your study perfectly.

migration(
  tz = NULL,
  sections,
  section.order = NULL,
  datapack = NULL,
  success.arrays = NULL,
  max.interval = 60,
  minimum.detections = 2,
  start.time = NULL,
  stop.time = NULL,
  speed.method = c("last to first", "last to last"),
  speed.warning = NULL,
  speed.error = NULL,
  jump.warning = 2,
  jump.error = 3,
  inactive.warning = NULL,
  inactive.error = NULL,
  exclude.tags = NULL,
  override = NULL,
  report = FALSE,
  auto.open = TRUE,
  discard.orphans = FALSE,
  discard.first = NULL,
  save.detections = FALSE,
  if.last.skip.section = TRUE,
  replicates = NULL,
  disregard.parallels = TRUE,
  GUI = c("needed", "always", "never"),
  print.releases = TRUE,
  plot.detections.by = c("stations", "arrays")
)

Arguments

tz
The time zone of the study area. Must match one of the values present in timezones.

sections
DEPRECATED - list the sections in the spatial input instead.

section.order
A vector containing the order by which sections should be aligned in the results.

datapack
A data bundle pre-compiled through the function preload. May be used to run actel analyses based on R objects, rather than input files.

success.arrays
The arrays that mark the end of the study area. If a tag crosses one of these arrays, the respective animal is considered to have successfully migrated through the study area.

max.interval
The number of minutes that must pass between detections for a new event to be created. Defaults to 60.

minimum.detections
For tags with only one movement event, defines the minimum number of times a tag must have been recorded during the study period for it to be considered true detections and not random noise. Defaults to 2.

start.time
Detection data prior to the timestamp set in start.time (in YYYY-MM-DD HH:MM:SS format) is not considered during the analysis.

stop.time
Detection data posterior to the timestamp set in stop.time (in YYYY-MM-DD HH:MM:SS format) is not considered during the analysis.

speed.method
Can take two forms: 'last to first' or 'last to last'. If 'last to first' (default), the last detection on the previous array is matched to the first detection on the target array to perform the calculations. If 'last to last', the last detection on ´the previous array is matched to the last detection on the target array to perform the calculations.

speed.warning
If a tag moves at a speed equal or greater than speed.warning (in metres per second), a warning is issued. If left NULL (default), no warnings are issued.

speed.error
If a tag moves at a speed equal or greater than speed.error (in metres per second), user intervention is suggested. If left NULL (default), user intervention is never suggested.

jump.warning
If a tag crosses a number of arrays equal or greater than jump.error without being detected, a warning is issued. If left NULL (default), no warnings are issued.

jump.error
If a tag crosses a number of arrays equal or greater than jump.error without being detected, user intervention is suggested. If left NULL (default), user intervention is never suggested.

inactive.warning
If a tag spends a number of days equal or greater than inactive.error in a given array at the tail of the respective detections, a warning is issued. If left NULL (default), no warnings are issued.

inactive.error
If a tag spends a number of days equal or greater than inactive.error in a given array at the tail of the respective detections, user intervention is suggested. If left NULL (default), user intervention is never suggested.

exclude.tags
A vector of tags that should be excluded from the detection data before any analyses are performed. Intended to be used if stray tags from a different code space but with the same signal as a target tag are detected in the study area.

override
A vector of signals for which the user intends to manually define which movement events are valid and invalid.

report
Logical. Should an HTML report be created at the end of the analysis? NOTE: Setting report to TRUE will generate an HTML file in the current directory. Additionally, if auto.open = TRUE (default), the web browser will automatically be launched to open the report once the function terminates.

auto.open
Logical: Should the report be automatically opened once the analysis is over? Defaults to TRUE. NOTE: If report = TRUE and auto.open = TRUE, the web browser will automatically be launched to open the report once the function terminates.

discard.orphans
Logical: Should actel automatically discard detections that do not fall within receiver deployment periods, or that were recorded before the respective animals were released?

discard.first
A threshold amount of time (in hours) that must pass after release for the respective detections to be valid. Set to 0 to discard only the release-to-first-detection calculations.

save.detections
Logical: Should the processed detections be saved for future runs?

if.last.skip.section
Logical: Should a tag detected at the last array of a given section be considered to have disappeared in the next section?

replicates
A list containing, for each array to which intra-array efficiency is to be calculated: The standard names of the stations to be used as a replicate. See the vignettes for more details.

disregard.parallels
Logical: Should the presence of parallel arrays invalidate potential efficiency peers? See the vignettes for more details.

GUI
One of "needed", "always" or "never". If "needed", a new window is opened to inspect the movements only when the movements table is too big to be displayed in R's console. If "always", a graphical interface is always created when the possibility to invalidate events emerges. If "never", a graphical interface is never invoked. In this case, if the table to be displayed does not fit in R's console, a temporary file will be saved and the user will be prompted to open that file and examine it. Defaults to "needed".

print.releases
Logical: Should the release sites be printed in the study area diagrams?

plot.detections.by
The type of y axis desired for the individual detection plots. One of "stations" (default) or "arrays".

Value

A list containing:

detections: A list containing all detections for each target tag;

valid.detections: A list containing the valid detections for each target tag;

spatial: A list containing the spatial information used during the analysis;

deployments: A data frame containing the deployments of each receiver;

arrays: A list containing the array details used during the analysis;

movements: A list containing all movement events for each target tag;

valid.movements: A list containing the valid movement events for each target tag;

section.movements: A list containing the valid section shifts for each target tag;

status.df: A data.frame containing summary information for each tag, including the following columns:

Times.entered.[section]: Number of times the tag was recorded entering a given section.

Average.time.until.[section]: Time spent between release or leaving another section and reaching at the given section.

Average.speed.to.[section]: Average speed from release or leaving one section and reaching the given section (if speed.method = "last to first"), or from release/leaving one section and leaving the given section (if speed.method = "last to last").

First.array.[section]: Array in which the tag was first detected in a given section

First.station.[section]: Standard name of the first station where the tag was detected in a given section

First.arrived.[section]: Very first arrival time at a given section

Average.time.in.[section]: Average time spent within a given section at each stay.

Average.speed.in.[section]: Average speed within a given section at each stay (only displayed if speed.method = "last to first").

Last.array.[section]: Array in which the tag was last detected in a given section

Last.station.[section]: Standard name of the last station where the tag was detected in a given section

Last.left.[section]: Very last departure time from a given section

Total.time.in[section]: Total time spent in a given section

Very.last.array: Last array where the tag was detected

Status: Fate assigned to the tag

Valid.detections: Number of valid detections

Invalid.detections: Number of invalid detections

Backwards.movements: Number of backward movement events

Max.cons.back.moves: Longest successive backwards movements

P.type: Type of processing:

'Skipped' if no data was found for the tag,

'Auto' if no user interaction was required,

'Manual' if user interaction was suggested and the user made changes to the validity of the events,

'Overridden' if the user listed the tag in the override argument.

Comments: Comments left by the user during the analysis

section.overview: A data frame containing the number of tags that disappeared in each section;

group.overview: A list containing the number of known and estimated tags to have passed through each array, divided by group;

release.overview: A list containing the number of known and estimated tags to have passed through each array, divided by group and release sites;

matrices: A list of CJS matrices used for the efficiency calculations;

overall.CJS: A list of CJS results of the inter-array CJS calculations;

intra.array.CJS: A list of CJS results of the intra-array CJS calculations;

times: A data frame containing all arrival times (per tag) at each array;

rsp.info: A list containing appendix information for the RSP package;

dist.mat: The distance matrix used in the analysis (if a valid distance matrix was supplied)

See also

Examples

# \donttest{
# Start by moving to a temporary directory
old.wd <- getwd()
setwd(tempdir())

# Deploy the example workspace
exampleWorkspace("exampleWorkspace")

#> M: The example workspace is now ready. To run the analysis on the example data, run:
#> 
#>   # move into the newly created folder
#>   setwd('exampleWorkspace')
#> 
#>   # Run analysis. Note: This will open an analysis report on your web browser.
#>   results <- explore(tz = 'Europe/Copenhagen', report = TRUE)
#> 
#> Once finished, explore the html report and the object 'results' for the output.

# Move your R session into the example workspace
setwd("exampleWorkspace")

# run the migration analysis. Ensure the tz argument
# matches the time zone of the study area and that the
# sections match your array names. The line below works
# for the example data.
results <- migration(tz = "Europe/Copenhagen")

#> M: Importing data. This process may take a while.

#> M: Number of target tags: 60.

#> M: Compiling detections...

#> M: Number of ALS: 17 (of which 1 had no detections)

#> Warning: No detections were found for receiver(s) 132907.

#> M: Data time range: 2018-04-13 22:26:37 to 2018-05-25 17:46:13 (Europe/Copenhagen).

#> M: Data successfully imported!

#> Warning: 'success.arrays' was not defined. Assuming success if the tags are last detected at array A9.

#> M: Creating movement records for the valid tags.

#> M: Checking movement events quality.

#> Warning: 'speed.warning'/'speed.error' were not set, skipping speed checks.

#> Warning: 'inactive.warning'/'inactive.error' were not set, skipping inactivity checks.

#> M: Compiling and checking section movements for the valid tags.

#> Warning: Inter-section backwards movements were detected for tag R64K-4508.

#> Warning: Inter-section backwards movements were detected for tag R64K-4526.

#> M: Filtering valid array and section movements.

#> M: Filling in the timetable.

#> M: Timetable successfully filled. Fitting in the remaining variables.

#> M: Getting summary information tables.

#> M: Validating detections...

#> Warning: No tags passed through array A0. Skipping efficiency estimations for this array.

#> M: Calculating array efficiency.

#> M: Skipping saving of the results.

#> M: Process finished successfully.

# to obtain an HTML report, run the analysis with report = TRUE

# return to original working directory
setwd(old.wd)
rm(old.wd)
# }