Calculate Transition Layer

Using a previously imported shape file that has been converted to a raster (see loadShape), Prepares a TransitionLayer object to be used in distance estimations (see distancesMatrix). Adapted from Grant Adams' script "distance to closest mpa".

transitionLayer(x, directions = c(16, 8, 4))

Arguments

x
A water raster; for example the output of loadShape

directions
The number of directions considered for every movement situation during cost calculation. See the vignettes for more details.

Value

A TransitionLayer object.

Details

It is highly recommended to read the vignette regarding distances matrix before running this function. You can find it by running vignette('a-2_distances_matrix', 'actel') or browseVignettes('actel')

Examples

# \donttest{
# check if R can run the distance functions
aux <- c(
  length(suppressWarnings(packageDescription("raster"))),
  length(suppressWarnings(packageDescription("gdistance"))),
  length(suppressWarnings(packageDescription("sp"))),
  length(suppressWarnings(packageDescription("tools"))),
  length(suppressWarnings(packageDescription("rgdal"))))
missing.packages <- sapply(aux, function(x) x == 1)

if (any(missing.packages)) {
  message("Sorry, this function requires packages '",
    paste(c("raster", "gdistance", "sp", "tools", "rgdal")[missing.packages], collapse = "', '"),
    "' to operate. Please install ", ifelse(sum(missing.packages) > 1, "them", "it"),
    " before proceeding.")
} else {
  if (suppressWarnings(require("rgdal"))) {
    # Fetch actel's example shapefile location
    aux <- system.file(package = "actel")[1]

    # import the shape file
    x <- loadShape(path = aux, shape = "example_shapefile.shp", size = 20)

    # Build the transition layer
    t.layer <- transitionLayer(x)

    # inspect the output
    t.layer

  } else {
    message("Sorry, it appears that rgdal is not being able to load.")
  }
}

#> M: Applying a small correction to the shape extent to ensure an integer number of pixels.

#> M: New shape extent:

#>         min       max
#> x  452961.4  453801.4
#> y 6242287.9 6242967.9

#> M: Chosen pixel size: 20 
#> M: Resulting pixel dimensions:

#>  x  y 
#> 42 34 

#> M: Burning the shape into a raster. This process may take several minutes depending on the shape size and chosen pixel size.

#> M: Constructing the transition layer. This process may take several minutes depending on the study area size and chosen pixel size.

#> class      : TransitionLayer 
#> dimensions : 34, 42, 1428  (nrow, ncol, ncell)
#> resolution : 20, 20  (x, y)
#> extent     : 452961.4, 453801.4, 6242288, 6242968  (xmin, xmax, ymin, ymax)
#> crs        : +proj=utm +zone=32 +datum=WGS84 +units=m +no_defs 
#> values      : conductance 
#> matrix class: dsCMatrix 

rm(aux, missing.packages)
# }